+++ src/projects/Project.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+interface Props {}
+
+const Project: React.FC = (props: Props) => {
+  return (
+    <div>
+      <h1>This is a project</h1>
+      <h1>
+        https://drive.google.com/file/d/1w0nsRuOq1uYGyq_UrIHI6YC2jQG2MPTb/view?usp=sharing
+      </h1>
+    </div>
+  );
+};
+
+export default Project;
